Remove implementations for `IntoIterator`, `FromIterator`, and `Extend`
It was brought to my attention that the `FromIterator` and `Extend`
implementations have no value to exist.

`TypeMap` maintains the invariant that for a given `T: TypeMapKey`
exists a `T::Value`, which is stored in a boxed trait object. These
implementations, however, allowed anyone to put any kind of value into
the boxed trait object, as well as insert a `TypeId` from any kind of
type, creating undefined behaviour due to violating the invariant.

An alternative to removing them would have been to accept `Box<S>`s, but
without the associated `T: TypeMapKey` type from which the `Box<S>` was
created from, it is not possible.

This also removes `IntoIterator` as there is likely not point in it
existing without the former traits.